pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "zuultimate"
version = "0.1.0"
description = "Enterprise identity, vault, zero-trust, and AI security platform"
readme = "README.md"
license = "MIT"
requires-python = ">=3.11"
authors = [{ name = "Chris Arseno" }]
keywords = ["security", "ai-security", "zero-trust", "vault", "identity"]

dependencies = [
    "fastapi>=0.115.0",
    "uvicorn[standard]>=0.30.0",
    "pydantic>=2.9.0",
    "pydantic-settings>=2.5.0",
    "sqlalchemy[asyncio]>=2.0.35",
    "aiosqlite>=0.20.0",
    "argon2-cffi>=23.1.0",
    "cryptography>=43.0.0",
    "pyjwt>=2.9.0",
    "typer>=0.12.0",
    "rich>=13.9.0",
    "httpx>=0.27.0",
    "pyyaml>=6.0.0",
    "structlog>=24.4.0",
    "alembic>=1.13.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0.0",
    "pytest-asyncio>=0.24.0",
    "pytest-cov>=5.0.0",
    "httpx>=0.27.0",
    "ruff>=0.6.0",
]
postgres = ["asyncpg>=0.29.0"]
redis = ["redis[hiredis]>=5.1.0"]

[project.scripts]
zuul = "zuultimate.cli:app"

[tool.hatch.build.targets.wheel]
packages = ["src/zuultimate"]

[tool.pytest.ini_options]
testpaths = ["tests"]
asyncio_mode = "auto"
filterwarnings = ["ignore::DeprecationWarning"]

[tool.ruff]
target-version = "py311"
line-length = 100
